pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "osint-toolkit"
version = "1.0.0"
description = "A modern web interface for various OSINT tools and resources"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
authors = [
    { name = "ayxkaddd", email = "bladeec1ty@protonmail.com" }
]

keywords = ["osint", "cybersecurity", "investigation"]

dependencies = [
    "fastapi>=0.115.8",
    "starlette>=0.45.3",
    "uvicorn>=0.34.0",
    "jinja2>=3.1.5",
    "aiohttp>=3.11.12",
    "httpx>=0.28.1",
    "requests>=2.32.3",
    "python-jose>=3.5.0",
    "pyjwt>=2.10.1",
    "passlib>=1.7.4",
    "bcrypt==4.3.0",
    "cryptography>=44.0.1",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.20",
    "pydantic>=2.10.6",
    "pydantic-settings>=2.10.1",
    "pydantic[email]>=2.10.6",
    "telethon>=1.39.0",
    "dnspython>=2.7.0",
    "python-whois>=0.9.5",
    "beautifulsoup4>=4.12.3",
    "socid-extractor>=0.0.27",
    "sse-starlette>=2.2.1",
    "pillow>=12.1.0",
    "aiofiles>=25.1.0"
]

[project.urls]
Homepage = "https://github.com/ayxkaddd/Osint-ToolKit"
Source = "https://github.com/ayxkaddd/Osint-ToolKit"
Issues = "https://github.com/ayxkaddd/Osint-ToolKit/issues"

[tool.hatch.build.targets.wheel]
packages = ["src"]